Create a separated file for the tests of EmptyDefaultConstructor
